Reorganized the code.
Split the UDPStream out into its own file.
Moved emitter & the stream files into `lib/`.
Moved examples into a new `examples/` directory.
Updated the lint & coverage targets to match.
