+++ examples/socket-emitter.js
