+++ examples/tcp-emitter.js
